Extract AppHeader to its own file
