+++ pages/_app.tsx
@@ -6,31 +6,17 @@ import {
   AlertDescription,
   AlertIcon,
   AlertTitle,
-  Button,
   ChakraProvider,
   Flex,
   Grid,
-  HStack,
-  Heading,
-  IconButton,
-  Image,
   Link,
-  Menu,
-  MenuButton,
-  MenuGroup,
-  MenuItem,
-  MenuList,
-  useColorMode,
-  useDisclosure,
 } from "@chakra-ui/react";
 import "../styles/globals.css";
 import "../styles/date-picker.css";
 import { theme } from "../styles/theme";
-import { useSessionUser } from "../session/useSessionUser";
-import { AuthModal } from "../components/Auth";
-import { SignUpModal } from "../components/SignUpModal";
 import React from "react";
-import { ExternalLinkIcon, MoonIcon, SunIcon } from "@chakra-ui/icons";
+import { ExternalLinkIcon } from "@chakra-ui/icons";
+import { AppHeader } from "../components/AppHeader";
 
 export default function MyApp({ Component, pageProps }) {
   return (
@@ -65,75 +51,5 @@ export default function MyApp({ Component, pageProps }) {
         </Grid>
       </ChakraProvider>
     </Authentication>
-  );
-}
-
-function AppHeader() {
-  const { session, logout } = useSessionUser();
-  const { isOpen: isAuthModalOpen, onOpen, onClose } = useDisclosure();
-  const {
-    isOpen: isSignUpModalOpen,
-    onOpen: onOpenSignUp,
-    onClose: onCloseSignUp,
-  } = useDisclosure();
-
-  return (
-    <>
-      <Flex as="header" align="center" justify="space-between" wrap="wrap">
-        <Link title="Home" href="/">
-          <Flex align="center" justify="flex-start">
-            <Image
-              m={1}
-              boxSize={["42px", "64px", "88px"]}
-              objectFit="cover"
-              src="/ssLogo.png"
-              alt="Shark Stewards"
-            />
-            <Heading m={1} fontSize={["large", "3xl", "3xl"]}>
-              Shark Watch CA
-            </Heading>
-          </Flex>
-        </Link>
-
-        <HStack spacing={2}>
-          <ThemeToggleButton />
-          <Menu>
-            <MenuButton variant="solid" m={3} as={Button}>
-              {session == null ? "Menu" : session.user.email}
-            </MenuButton>
-            <MenuList>
-              {session == null && <MenuItem onClick={onOpen}>Login</MenuItem>}
-              {session == null && (
-                <MenuItem onClick={onOpenSignUp}>Sign Up</MenuItem>
-              )}
-              {session != null && (
-                <>
-                  <MenuItem>My Sightings</MenuItem>
-                  <MenuGroup>
-                    <MenuItem>Profile</MenuItem>
-                    <MenuItem onClick={logout}>Sign Out</MenuItem>
-                  </MenuGroup>
-                </>
-              )}
-            </MenuList>
-          </Menu>
-        </HStack>
-      </Flex>
-      {isAuthModalOpen && <AuthModal onClose={onClose} />}
-      {isSignUpModalOpen && <SignUpModal onCloseSignUp={onCloseSignUp} />}
-    </>
-  );
-}
-
-function ThemeToggleButton() {
-  const { colorMode, toggleColorMode } = useColorMode();
-  const Icon = colorMode === "light" ? MoonIcon : SunIcon;
-  return (
-    <IconButton
-      aria-label="Toggle Theme"
-      variant="ghost"
-      onClick={toggleColorMode}
-      icon={<Icon />}
-    />
   );
 }
